Requirement 2 Complete.
When you click one the thumbnail image or the name of the movie, on the listings page, it should either open a modal window that shows you more information about the movie you clicked on, or take you to a new page that shows you more information about the movie you clicked on. This can be “ipsum dolor” placeholder text, but should vary from movie to movie.