Requirement 3 Complete.
From the listings page, there should be a button to either show a modal window or take you to a new page, where you can create a new movie listing. It should at least collect the movie title, a description, and allow you to either upload an image/thumbnail for the movie, or let you enter a direct image url.