Fixed "NotFound" and "Error" view not found errors.
Had to add the "_Layout" as the master file name.
Added more Actors in the database seeder.
Trimming the Movie Title and Description when creating a listing.
Only showing the first 500 characters of Movie Description when viewing the Movie Listings. Full Description can be viewed by clicking on "More" (previously "Details")